+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 - Bugfix for problem with list_documents() when no documents present in store
+- The append-only property of tables created by `pw.io.kafka.read` is now set correctly.
 
 
 ## [0.21.1] - 2025-03-28